rlookup
rlookup is a function which searches for matches between objects (coerced to matrix) within a range of values (not exact match). The data from one column in object 1 is searched for in object 2 and any match that is within the specified range is counted as a hit. This function is analogous to but more versatile than Microsoft Excel VLOOKUP function.