mzunpack
mzunpack is designed to convert a specifically structured list object created by the peaks() function in the mzR package, where every entry in the list is a two-column matrix: column 1 contains mass-to-charge (m/z) values and column 2 contains the peak intensity registered for that mass channel. Each matrix entry in the list object represents a single mass spectrum scan. The time values for the scans can be extracted from the open-source format using the header() function in the package mzR.