regress.all
Performs linear regression (using lm()) for all n-point groups (specified by "points" argument) in a matrix above a certain R^2 and slope cutoff. 

A note about regress.all()
When performing this function on indexed matrices, make sure the MZ index results in a single vector 
of m/z values, while the identically indexed RET matrix returns a vector of retention times whose 
values correspond exactly to the values in the MZ index vector. A useful way to achieve this is to have the m/z values as a column in the data matrix and reference column 1 as follows: regress.all(datamatrix[,1][which(datamatrix... >cutoff,arr.in=T)[,1]] ... ). The which() indexer collects the values in datamatrix which obey the cutoff (“hits”), the [,1] after the which() command ensures that the row number in the datamatrix where the hit was found is collected, and the [,1] on the far left in bold indicates that only the first value in the row containing the hit will be displayed (the first value here is the m/z value). 
An error that can exist is the following:
Error in lm.fit(x, y, offset = offset, singular.ok = singular.ok, ...) : 
0 (non-NA) cases
This error has its origin from the lm() function within regress.all(). The error occurs because the lm() function has produced NAs. The most common error to produce this is not providing the correct indeces. The input for regress.all() can be confusing. For the following data window:
datamatrix[-1:-1100,c(-1:-750,-850:-1860)]

The input for regress.all() using this window would look like the following:

regress.all.matrix(datamatrix[c(-1:-1100),1][which(datamatrix[c(-1:-1100),c(-1:-750,-850:-1860)]>8,arr.in=T)[,1]],retention_time_vector[c(-1:-749,-849:-1859)][which(datamatrix[c(-1:-1100),c(-1:-750,-850:-1860)]>8,arr.in=T)[,2]],2,0.999)

1.	The first input is datamatrix[c(-1:-1100),1][which(datamatrix[c(-1:-1100),c(-1:-750,-850:-1860)]>8,arr.in=T)[,1]]. 
2.	The second input is retention_time_vector[c(-1:-749,-849:-1859)][which(datamatrix[c(-1:-1100),c(-1:-750,-850:-1860)]>8,arr.in=T)[,2]]. 
3.	The third input is 2.
4.	The fourth input is 0.999.