abline.batch
Plots least-squares fit lines based on results from regress.all.

Example:
abline.batch(reg3[which(reg3[  ,1]==713.5 & reg3[  ,7]>0),  ], 7, 8)
1.	abline.batch() is the function call
2.	reg3 is the name of the data matrix
3.	reg3[which(reg3[,1]==713.5 & reg3[,7]>0), ] is consists of two search conditions on reg3:
a.	which(reg3[,1]==713.5 finds the entries of m/z 713.5 in column 1 of reg3. Column 1 of reg3 consists of numbers that were part of the regression function.
b.	The single & is to add another search condition. The double &&, which is not applicable in this case, is a logical operator.
c.	reg3[,7]>0),  ] refines the first search term (described in point 3a) to search for those regressions which include the point 713.5 and have a positive slope, since slope is in the 7th column of matrix reg3. The ,  ] means that the result of this regression will include all columns of reg3 in the indexed result.
d.	The final numbers 7 and 8 refer to the